+++ src/clients/index.js
@@ -1,4 +1,3 @@
-export * from './db.js'
 export * from './client-get-order.js'
 export * from './client-catalog.js'
 export * from './client-hook.js'
