+++ src/presentation/controllers/CombinationController.js
@@ -1,5 +1,4 @@
 import axios from 'axios';
-import { GetOrder } from '../../clients/client-get-order.js';
 import mergeSort from '../../services/protocols/MergeSort.js';
 export default new class CombinationController {
