+++ src/presentation/controllers/controller-top-combinations.js
@@ -1,4 +1,4 @@
-import {ServicesCombinations} from '../../app/services/services-combinations.js'
+import {ServicesCombinations} from '../../services/services-combinations.js'
 
 export class ControllerTopCombinations {
   static async handle (req, res) {
